requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-nested-routers==0.94.1
dj-rest-auth==7.0.1
django-allauth==65.3.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-storages==1.14.4
boto3==1.35.80
requests==2.32.3
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
whitenoise==6.8.2
